+++ test/unit/statement_test.exs
@@ -1,0 +1,6 @@
+defmodule RDF.StatementTest do
+  use RDF.Test.Case
+
+  doctest RDF.Statement
+
+end